+++ internal/query/execute_query.go
@@ -31,6 +31,7 @@ type executeSettings interface {
 	CallOptions() []grpc.CallOption
 	RetryOpts() []retry.Option
 	ResourcePool() string
+	ResponsePartLimitSizeBytes() int64
 }
 
 type executeScriptConfig interface {
@@ -77,6 +78,7 @@ func executeQueryRequest(a *allocator.Allocator, sessionID, q string, cfg execut
 	request.StatsMode = Ydb_Query.StatsMode(cfg.StatsMode())
 	request.ConcurrentResultSets = false
 	request.PoolId = cfg.ResourcePool()
+	request.ResponsePartLimitBytes = cfg.ResponsePartLimitSizeBytes()
 
 	return request, cfg.CallOptions()
 }
@@ -146,14 +148,7 @@ func readResultSet(ctx context.Context, r *streamResult) (_ *resultSetWithClose,
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
-
-	_, err = r.nextResultSet(ctx)
-	if err == nil {
-		return nil, xerrors.WithStackTrace(errMoreThanOneResultSet)
-	}
-	if !xerrors.Is(err, io.EOF) {
-		return nil, xerrors.WithStackTrace(err)
-	}
+	rs.mustBeLastResultSet = true
 
 	return &resultSetWithClose{
 		resultSet: rs,